Sync with latest Fast Downward and remove state-dependent action costs
* [trivial] Update contributor info.

* [issue992] Make landmark code more readable.

Refactor all files in `src/search/landmarks` with the main objective to
structure them to clarify what they do. The main changes include
 - breaking apart long functions,
 - renaming variables and functions, and
 - making code consistent with our coding conventions.
Further changes include
 - using more suitable data structures,
 - improving simple procedures, and
 - using the `ranges` library.

The general functionality remains unchanged, or at least that was our
intention. We verified experimentally that this appears to be the case
for the most commonly used landmark configurations. Performance was
affected positively over all, mainly because landmark generation is
faster with most factories. Most remarkably, the h^m landmark factory
solves 24 more problems, but also LAMA improves slightly in terms of
coverage (3-4 more problems solved).

Co-authored-by: Remo Christen <remo.christen@unibas.ch>
Co-authored-by: Florian Pommerening <florian.pommerening@unibas.ch>

* [trivial] Skip delete action if artifact is missing.

* [trivial] Fix typo in SCC acronym and expand it in synopsis.

* [issue 1163] Print code that was used to build the code

We now print the code revision that was used to build the search code.
This is printed to to the log at the start of every planner run and 
also be queried directly with `./fast-downward.py --version`.
As part of this change, we dropped support for builds outside of the
`builds` directory.

* [issue1184] translator: Add --keep-no-ops option

We added a new option `--keep-no-ops` to the translator, which defaults
to false.

To enable the option, add `--translate-options --keep-no-ops` to the end
of the argument list for `fast-downward.py`.

The translator component usually prunes all ground actions that have no
effects. When the new option is set to true, these actions are no longer
pruned. There is generally no good reason to use this option in regular
classical planning settings. It is primarily intended for users that extend
Fast Downward to richer planning formalisms such as FOND planning or
other scenarios where the planner is part of a larger pipeline.

On the extended downward-benchmark benchmark set, 1208 out of 3955
tasks for which we could finish translation within the resource limits
in our experiments produce at least one action with no effect with the
new flag.

* [trivial] README.md: Update contributors list for issue1184.

* [trivial] README.md: update general copyright year; rewrap long paragraphs

* [issue1188] Add a script that creates a local build of the website.

The script is located in misc/website and combines the generated
documentation, the documentation from docs/ and the one from
the separate downward-markdown repository.

* [issue1186] Refactor the EagerSearch step function without changing the search behavior.

- Split logic from step() into get_next_node_to_expand() and expand().
- Move successor generation to a dedicated generate_successors() method.
- Add helper functions for collecting preferred operators.

* [issue1189] Check code style with clang-format instead of uncrustify.

Because uncrustify does currently not correctly handle modern C++ features such as concepts, we switch to the more up-to-date clang-format. This makes our style checks more opinionated and more strict. Our configuration is based on the LLVM style and defined in `.clang-format`. To reformat all source files according to the configuration you can for example run the following command in the repository root: `clang-format -i src/search/**/*.{cc,h}`.

* [issue1146] Validate SAS file (#227)

The search component now does more checking of the validity of its input file.
We add a more thorough diagnostic message when invalid input is provided. We
restructured to explicitly split the parser, lexer and internal task 
representation. The SAS task parsing now takes roughly 25% more time in our
experiments.

Co-authored-by: Claudia Grundke <claudia.grundke@unibas.ch>
Co-authored-by: Florian Pommerening <florian.pommerening@unibas.ch>
Co-authored-by: Salomé Eriksson <salome.eriksson@unibas.ch>
Co-authored-by: Malte Helmert <malte.helmert@unibas.ch>
Co-authored-by: Travis Rivera Petit <bedowu54@dmi-24-pc-16.dmi.unibas.ch>

* [trivial] specify clang-format version. (#274)

* [issue1185] Make the translator a python package.

The translator is now a python package (within Fast Downward).
This does not affect how Fast Downward is called, but if you want to
call the translator directly then use 'python3 -m translate' instead of
'python3 translate/main.py' (which replaced the previous 'translate.py').

* [trivial] Fix accessing a translator option.

* [issue1190] Update documentation on the Fast Downward website.

The online documentation for Fast Downward has been updated. 
Especially the steps for getting started with Fast Downward and running
the planner should be clearer now.

* [trivial] README.md: Update contributors list.

* [issue1192] Fix default value handling in MIASM scoring function.

The MIASM scoring function now handles default/missing values (-1)
for parameters max_states, max_states_before_merge, and
shrink_threshold_before_merge in the same way as the merge-and-shrink
heuristic does.

* [trivial] Replace deprecated pipes.quote() with shlex.quote().

* [issue1191] Refactor interface of MergeSelector.

MergeSelector now has two variants of the method for choosing the best
merge candidate for a given factored transition system: one that
considers all merge candidates for the given factored transition
system, and another one that only considers the given (sub)set of merge
candidates.

* New style

* rm clang-tidy from style github action

* Fix forward declaration issue with MutexGroup

* Update Readme and fix empty op error in sas parsing

* Fix style

---------

Co-authored-by: Malte Helmert <malte.helmert@unibas.ch>
Co-authored-by: Clemens Büchner <clemens.buechner@unibas.ch>
Co-authored-by: Remo Christen <remo.christen@unibas.ch>
Co-authored-by: Florian Pommerening <florian.pommerening@unibas.ch>
Co-authored-by: Simon Dold <simon.dold@stud.unibas.ch>
Co-authored-by: Masataro Asai <guicho271828@users.noreply.github.com>
Co-authored-by: Christian Muise <christian.muise@gmail.com>
Co-authored-by: Gabi Roeger <gabriele.roeger@unibas.ch>
Co-authored-by: Tanja Schindler <tanja.schindler@unibas.ch>
Co-authored-by: Claudia Grundke <claudia.grundke@unibas.ch>
Co-authored-by: Salomé Eriksson <salome.eriksson@unibas.ch>
Co-authored-by: Travis Rivera Petit <bedowu54@dmi-24-pc-16.dmi.unibas.ch>
Co-authored-by: SimonDold <48084373+SimonDold@users.noreply.github.com>
Co-authored-by: Jendrik Seipp <jendrik.seipp@liu.se>
Co-authored-by: Esther Mugdan <71381124+emugdan@users.noreply.github.com>
Co-authored-by: Silvan Sievers <silvan.sievers@unibas.ch>